add WorkerProvider interface
